+++ ReleaseNotes.md
@@ -90,6 +90,7 @@
 ### vX.X.X (upcoming release)
 #### Features
 * New control `StepTitleHeaderControl` to use bindings in a XAML defined step
+* `TabControl` styles to render it in [Material Design](https://material.io/design/components/tabs.html)
 #### Fixes
 * Handle broken image files in `BitmapImageHelper`
 * Handle issues if MaterialDesignThemes version 2.5.0.1205 is used